abstract to HoodieBaseTable
